Connection Timeout for Health Check Requests
Under certain conditions it is possible for health check requests to never return. This commit adds a connection timeout to the http client in order to prevent this from bringing down the entire health exporter. The problem can be easily reproduced by adding a non-routable IP (like 10.255.255.1) as health check target to the configuration.